WebSocketClientSocketIO.ino:113:25: ArduinoJson, you must set ARDUINOJSON_USE_LONG_LONG error

+++ examples/esp8266/WebSocketClientSocketIO/WebSocketClientSocketIO.ino
@@ -110,7 +110,7 @@ void loop() {
 
         // add payload (parameters) for the event
         JsonObject param1 = array.createNestedObject();
-        param1["now"] = now;
+        param1["now"] = (uint32_t) now;
 
         // JSON to String (serializion)
         String output;
